creates lintworker helpers file
