+++ packages/lint-worker/package.json
@@ -31,25 +31,31 @@
   },
   "dependencies": {
     "@neo4j-cypher/language-support": "workspace:*",
-    "workerpool": "^9.0.4"
+    "languageSupport-next.13": "npm:@neo4j-cypher/language-support@2.0.0-next.13",
+    "languageSupport-next.3": "npm:@neo4j-cypher/language-support@2.0.0-next.3",
+    "@neo4j-cypher/query-tools": "workspace:*",
+    "workerpool": "^9.0.4",
+    "vscode-languageserver": "^8.1.0"
   },
   "scripts": {
-    "build": "pnpm build-types && pnpm build-esm && pnpm build-commonjs",
-    "build-commonjs": "esbuild ./src/lintWorker.ts --bundle --platform=node --conditions=require --outfile=dist/cjs/lintWorker.cjs --format=cjs --minify",
-    "build-esm": "esbuild ./src/lintWorker.ts --bundle --platform=node --conditions=require --outfile=dist/esm/lintWorker.mjs --format=esm --minify",
+    "build": "pnpm build-types && pnpm build-lintworker-esm && pnpm build-lintworker-commonjs && pnpm build-commonjs && pnpm build-esm",
+    "build-lintworker-commonjs": "esbuild ./src/lintWorker.ts --bundle --platform=node --conditions=require --outfile=dist/cjs/lintWorker.cjs --format=cjs --minify",
+    "build-lintworker-esm": "esbuild ./src/lintWorker.ts --bundle --platform=node --conditions=require --outfile=dist/esm/lintWorker.mjs --format=esm --minify",
+    "build-commonjs": "esbuild ./src/index.ts --bundle --platform=node --conditions=require --outfile=dist/cjs/index.cjs --format=cjs --minify",
+    "build-esm": "esbuild ./src/index.ts --bundle --platform=node --conditions=require --outfile=dist/esm/index.mjs --format=esm --minify",
     "clean": "rm -rf dist",
     "build-types": "tsc --emitDeclarationOnly --outDir dist/types"
   },
   "devDependencies": {
     "@types/lodash.debounce": "^4.0.9"
   },
-  "types": "./dist/types/src/lintWorker.d.ts",
+  "types": "./dist/types/src/index.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/types/src/lintWorker.d.ts",
-      "require": "./dist/cjs/lintWorker.cjs",
-      "import": "./dist/esm/lintWorker.mjs",
-      "default": "./dist/esm/lintWorker.mjs"
+      "types": "./dist/esm/index.d.ts",
+      "require": "./dist/cjs/index.cjs",
+      "import": "./dist/esm/index.mjs",
+      "default": "./dist/esm/index.mjs"
     }
   }
 }
