+++ packages/lint-worker/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lintWorker';
+export * from './helpers';
